repos/rental-finder-ai/backend/requirements.txt

fastapi
uvicorn
openai
python-dotenv
requests
beautifulsoup4
